Use background task priorities
The priorities of background tasks will be used
to give the repo update (and publish) task a lower priority,
so that all other tasks that still modify the repository get executed first.

This should also reduce the number of required updates and publish tasks,
because before a new screenshot would be downloaded right after an
update and then trigger a new update.

Now all tasks that cause updates are executed before the actual update
happens, so only one update needs to be performed in the end.

This commit also lowers the priority of remote repo updates,
so that all other tasks are done first.