Fix app icons on repo page
We always had the icons of apps in the repo dir but this changed sometimes ago. The page still assumed that they were in the root.

This commit changes it to "icons-640/" and I have tested it with several different apps, both from remote repos and locally, because I can remember this somehow did not work when implementing the repo page.

Fixes #143.