Do not use app metadata overrides if it can be avoided
If metadata overrides are set in a remote app,
they will be moved into the default language.
If the information already exists in the default language,
the overrides will be used unchanged.
Only in this case, the override form fields will be shown
when the app is edited.
Otherwise, the override fields are not even visible to the user.