Make internal language-code handling case-insensitive
Language codes used by Django are all lower-case.
In order to be able to compare them properly,
this commit stores them lower-cased as well.

To stay compatible with fdroidserver and fdroidclient,
this commit uses the more common way of writing a language code
with the region part upper-cased when writing to the repo index.
Files are also stored on disk and made available via URLs
with the more common way of writing the code.

So all lower-case instances should be purely internal.
Tests have been added to ensure this.

Fixes #156