Fix endless scroll when all apps are added
Before, all app cards had only the class 'rm-app-card--no-hover' when their
app was already added to the repo. This caused endless scroll to fail
because it was looking for app cards with the class 'rm-app-card'.

Now, all app cards have the class 'rm-app-card' all the time. When an
app card should not change on hover, the class 'rm-app-card--no-hover'
is added so both 'rm-app-card' and 'rm-app-card--no-hover' exist.