Add uploaded files to page instead of reloading it
Every time the user uploaded files in app edit mode via drag and drop it
caused the site to refresh in order to show the files on it. This caused
all other files the user has entered information in, like description
and author, to be wiped because we currently have no auto-save
functionality here (#88).

Now, when users upload files via drag and drop, Repomaker returns
information about them in a JSON response which gets used by JavaScript
to dynamically add the content to the page.

Fixes #150.