Implement button to clear feature graphic in app edit
For now I implemented the deletion of feature graphics like the one of
screenshots and APKs: a separate page opens where the users confirm they
want to delete the feature graphic.

I did not implemented a special POST to app edit because this solution
seems cleaner and aligns well to what's known in Repomaker. In another
commit I will catch the delete events, show a JavaScript dialog and
delete the objects by posting with JavaScript in the background to the
known views.

Fixes #168.