Use same password for store and key
Some keystore types (e.g. PKCS12) don't support different passwords for
store and key. We don't really need different passwords for them,
therefore this commit changes the behavior of Repomaker when creating
repos to use the same password for both the store and the key.
The same way those passwords are handled by _fdroidserver_.

To support former versions of Repomaker, the passwords for store and key
are still handled differently, although starting with this commit they
are now the same.

Related to https://gitlab.com/fdroid/fdroid-repomaker-flatpak/issues/13.